refactor(sky-server): 重构代码并优化套餐相关功能
- 修改菜品批量删除接口名称，提高代码一致性
- 更新套餐相关接口，使用 Long 类型替代 Integer- 新增套餐删除功能的实现
- 优化套餐查询接口，改进代码结构
- 添加套餐与菜品关联查询的 SQL 语句

+++ sky-server/src/main/java/com/sky/controller/admin/setmealController.java
@@ -24,7 +24,7 @@ public class setmealController {
     
     @GetMapping("/{id}")
     @ApiOperation(value = "根据id查询套餐")
-    public Result<SetmealVO> getSetMealById(@PathVariable Integer id) {
+    public Result<SetmealVO> getSetMealById(@PathVariable Long id) {
         SetmealVO setmealVO = setMealService.getSetMealById(id);
         log.info("套餐id为{}的信息{}", id, setmealVO);
         return Result.success(setmealVO);
